requirements.txt
beautifulsoup4
requests
lxml
pillow
pyzbar
